Draw_Card, Place_Card, Minor Changes
Created 2 New Functions:

Draw_Card: Draws the card with the highest index with a Played value of false from UNOCardTable, switching that Played value to true, and adding that card to the hand of currentPlayer.

Place_Card: Obtains index of the previous card and selected card given their ID's, then checks if selected card can be placed. If the card can be placed, the card ID is replaced with a 0 in the player's hand, and the Played value of the selected card in UNOCardTable is switched back to false.

Minor Changes: fixed bug where Deck_Creator did not randomize every row, added semicolons in some other places.

I plan to continue working on Place_Card, potenitally adding functionality for checking the type of card and dealing with the consequences (wild, draw2, reverse, skip, etc.)